fix(entity/user): 유저의 스터디 횟수와 혼자연습 횟수 데이터타입 변경

+++ src/main/java/com/witherview/database/entity/User.java
@@ -39,10 +39,10 @@ public class User {
     private String subJob;
 
     @ColumnDefault("0")
-    private Integer studyCnt;
+    private Long studyCnt;
 
     @ColumnDefault("0")
-    private Integer selfPracticeCnt;
+    private Long selfPracticeCnt;
 
     @ColumnDefault("0")
     private Byte reliability;
